Fix StreamingTensorInfo.nBytes Int overflow for tensors > 2 GB
Change nBytes from Int to Long to prevent silent overflow when computing
byte sizes for large tensors (e.g., Gemma 4 E4B per_layer_token_embd.weight).

Root cause: `(numBlocks * typeSize).toInt()` silently truncated a Long
result exceeding Int.MAX_VALUE, producing negative byte counts that
caused `IllegalArgumentException: Length must be non-negative` in
JvmRandomAccessSource.readAt().

Changes:
- StreamingTensorInfo.nBytes: Int → Long
- StreamingGGUFReader: keep nBytes computation in Long, no .toInt()
- StreamingGGUFReader.loadTensorData: explicit check with clear error
  message for tensors > 2 GB (use loadTensorStorageMapped instead)
- GGUFReader (legacy): add overflow guard with actionable error message
- Update all test comparisons for Long nBytes

Fixes #452

+++ skainet-io/skainet-io-gguf/src/jvmTest/kotlin/sk/ainet/io/gguf/QuantizedModelLoadTest.kt
@@ -49,8 +49,8 @@ class QuantizedModelLoadTest {
 
                 // Verify nBytes calculation: (nElements / blockSize) * typeSize
                 val (blockSize, typeSize) = GGML_QUANT_SIZES[tensor.tensorType]!!
-                val expectedBytes = (tensor.nElements / blockSize) * typeSize
-                assertEquals(expectedBytes.toInt(), tensor.nBytes,
+                val expectedBytes = (tensor.nElements.toLong() / blockSize) * typeSize
+                assertEquals(expectedBytes, tensor.nBytes,
                     "Tensor ${tensor.name}: nBytes mismatch - expected $expectedBytes, got ${tensor.nBytes}")
             }
 
@@ -59,7 +59,7 @@ class QuantizedModelLoadTest {
             if (smallTensor != null) {
                 println("\nLoading smallest tensor: ${smallTensor.name} (${smallTensor.nBytes} bytes)")
                 val data = reader.loadTensorData(smallTensor)
-                assertEquals(smallTensor.nBytes, data.size,
+                assertEquals(smallTensor.nBytes, data.size.toLong(),
                     "Loaded data size should match nBytes")
                 println("Successfully loaded ${data.size} bytes!")
             }
